# local_eats_api_gateway/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/order_service/comment/create": {
            "post": {
                "description": "Create a new comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Create Comment",
                "parameters": [
                    {
                        "description": "Create Comment",
                        "name": "Create",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.CreateReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/genproto.ReviewResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/order_service/comment/{kitchen_id}": {
            "get": {
                "description": "Get reviews for a kitchen based on query parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "Get Reviews for Kitchen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kitchen ID",
                        "name": "kitchen_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/genproto.GetReviewsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/order_service/meal/{meal_id}/update-nutrition-info": {
            "put": {
                "description": "Updates the nutrition information of a dish based on the provided request payload.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qo'shimcha Api"
                ],
                "summary": "Update nutrition information of a dish",
                "operationId": "updateNutritionInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Meal ID",
                        "name": "meal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.UpdateNutritionInfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated dish information",
                        "schema": {
                            "$ref": "#/definitions/genproto.Dish"
                        }
                    }
                }
            }
        },
...
        "/api/user_service/kitchen/statistics/{kitchen_id}": {
            "get": {
                "description": "Retrieves statistics for a kitchen based on the provided kitchen_id, start_date, and end_date.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qo'shimcha Api"
                ],
                "summary": "Retrieve kitchen statistics",
                "operationId": "getKitchenStatistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kitchen ID",
                        "name": "kitchen_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistics for the kitchen",
                        "schema": {
                            "$ref": "#/definitions/genproto.KitchenStatisticsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user_service/kitchen/update-working-hours": {
            "put": {
                "description": "Updates the working hours of a kitchen based on the provided request payload.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qo'shimcha Api"
                ],
                "summary": "Update kitchen working hours",
                "operationId": "updateWorkingHours",
                "parameters": [
                    {
                        "description": "Request payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.UpdateWorkingHoursRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated working hours information",
                        "schema": {
                            "$ref": "#/definitions/genproto.UpdateWorkingHoursResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/user_service/users/activity/{user_id}": {
            "get": {
                "description": "Retrieves activity statistics for a user based on the provided start_date and end_date query parameters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qo'shimcha Api"
                ],
                "summary": "Retrieve user activity",
                "operationId": "getUserActivity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity statistics for the user",
                        "schema": {
                            "$ref": "#/definitions/genproto.GetUserActivityResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "genproto.BusiestHour": {
            "type": "object",
            "properties": {
                "hour": {
                    "type": "integer"
                },
                "orders_count": {
                    "type": "integer"
                }
            }
        },
...
        "genproto.CreateReviewRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "kitchen_id": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "genproto.CuisineActivity": {
            "type": "object",
            "properties": {
                "cuisine_type": {
                    "type": "string"
                },
                "orders_count": {
                    "type": "integer"
                }
            }
        },
        "genproto.Dish": {
            "type": "object",
            "properties": {
                "allergens": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "dietary_info": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nutrition_info": {
                    "$ref": "#/definitions/genproto.NutritionInfo"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "genproto.GetReviewsResponse": {
            "type": "object",
            "properties": {
                "average_rating": {
                    "type": "number"
                },
                "limit_offset": {
                    "$ref": "#/definitions/genproto.LimitOffset"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.ReviewResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "genproto.GetUserActivityResponse": {
            "type": "object",
            "properties": {
                "favorite_cuisines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.CuisineActivity"
                    }
                },
                "favorite_kitchens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.KitchenActivity"
                    }
                },
                "total_orders": {
                    "type": "integer"
                },
                "total_spent": {
                    "type": "number"
                }
            }
        },
        "genproto.KitchenActivity": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "orders_count": {
                    "type": "integer"
                }
            }
        },
...
        "genproto.KitchenStatisticsResponse": {
            "type": "object",
            "properties": {
                "average_rating": {
                    "type": "number"
                },
                "busiest_hours": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.BusiestHour"
                    }
                },
                "top_dishes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.TopDish"
                    }
                },
                "total_orders": {
                    "type": "integer"
                },
                "total_revenue": {
                    "type": "number"
                }
            }
        },
...
        "genproto.LimitOffset": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                }
            }
        },
...
        "genproto.NutritionInfo": {
            "type": "object",
            "properties": {
                "calories": {
                    "type": "number"
                },
                "carbohydrates": {
                    "type": "number"
                },
                "fat": {
                    "type": "number"
                },
                "protein": {
                    "type": "number"
                }
            }
        },
...
        "genproto.ReviewResponse": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kitchen_id": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "genproto.TimeRange": {
            "type": "object",
            "properties": {
                "close": {
                    "type": "string"
                },
                "open": {
                    "type": "string"
                }
            }
        },
        "genproto.TopDish": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "orders_count": {
                    "type": "integer"
                },
                "revenue": {
                    "type": "number"
                }
            }
        },
...
        "genproto.UpdateNutritionInfoRequest": {
            "type": "object",
            "properties": {
                "allergens": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "dietary_info": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "dish_id": {
                    "type": "string"
                },
                "nutrition_info": {
                    "$ref": "#/definitions/genproto.NutritionInfo"
                }
            }
        },
...
        "genproto.UpdateWorkingHoursRequest": {
            "type": "object",
            "properties": {
                "kitchen_id": {
                    "type": "string"
                },
                "working_hours": {
                    "$ref": "#/definitions/genproto.WorkingHours"
                }
            }
        },
        "genproto.UpdateWorkingHoursResponse": {
            "type": "object",
            "properties": {
                "kitchen_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "working_hours": {
                    "$ref": "#/definitions/genproto.WorkingHours"
                }
            }
        },
...
        "genproto.WorkingHours": {
            "type": "object",
            "properties": {
                "friday": {
                    "$ref": "#/definitions/genproto.TimeRange"
                },
                "monday": {
                    "$ref": "#/definitions/genproto.TimeRange"
                },
                "saturday": {
                    "$ref": "#/definitions/genproto.TimeRange"
                },
                "sunday": {
                    "$ref": "#/definitions/genproto.TimeRange"
                },
                "thursday": {
                    "$ref": "#/definitions/genproto.TimeRange"
                },
                "tuesday": {
                    "$ref": "#/definitions/genproto.TimeRange"
                },
                "wednesday": {
                    "$ref": "#/definitions/genproto.TimeRange"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}